# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "responses": {
          "200": { "description": "Basic API metadata" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "Service status" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing fields or Supabase rejected the signup" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Access token and refresh token" },
          "400": { "description": "Missing fields" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public, open info — no auth required",
        "responses": {
          "200": { "description": "Public message" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private profile data (bearer token required)",
        "responses": {
          "200": { "description": "Profile data" },
          "401": { "description": "Missing, malformed, or invalid access token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Example second protected route reusing the same auth guard",
        "responses": {
          "200": { "description": "Dashboard message" },
          "401": { "description": "Missing, malformed, or invalid access token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "End the user's session (bearer token required)",
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Missing, malformed, or invalid access token" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "responses": {
          "200": { "description": "An array of tasks" }
        }
      },
      "post": {
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "example": "Buy milk" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Invalid body" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get one task by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "The task" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task's title and/or done status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "minLength": 1 },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "The updated task" },
          "400": { "description": "Invalid or empty update" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}